feat: also run extension custom scripts for superusers
Run custom scripts even when the current role is a superuser. This lets
us do things that needs to happen unconditionally, e.g. granting
privileges to existing roles.

If we *do* want to run certain scripts only when the current role is not
a superuser, we can simply use a conditional like `if (select rolsuper from pg_roles where rolname = ...) then ...`