feat: constrained extensions
Adds resource constraints for cpu, memory and disk for extensions.

Compat with pg 12 is done by omitting constrainted extensions code using
if macro directives checking the PostgreSQL version.

Since all the tests are ran for all the pg versions, a variant comparison
file for pg12 is used:

https://www.postgresql.org/docs/current/regress-variant.html

The variant test checks that all the extensions are created and
that this only happens for pg 12. This way we ensure that the pg12
test output is not matched by accident for pg >= 13 and we
get test failures if the constrained extensions code fails.