fix: switch to privileged_extensions_superuser even if superuser
For extensions like pg_cron, the permissions depend on the owner of the
extension object, which in turn depends on which role runs the `CREATE
EXTENSION`.

One might want to temporarily make the user role a superuser during
pg_restore. This would break pg_cron because once the user role is
reverted to non-superuser, the extension owner (i.e. the non-superuser
user role) would not have permissions to execute pg_cron functions.

This is not an issue if we make the `CREATE EXTENSION` run as the
privileged_extensions_superuser unconditionally, since by definition it
would always be a superuser.