chore: enable code coverage
Requires changing the development workflow. Installation and packaging works as usual.

Goes from this:

```
supautils-with-pg-14 make installcheck
```

To this:

```
nxpg-14 nxpg-build && nxpg-14 nxpg-tmp make installcheck
```

While it's more verbose, compiling is faster now because you don't
need to go inside and out of `nix-shell` each time you make a change on
a source file.

Going inside the tmp pg still works as usual:

```
nxpg-14 nxpg-tmp psql
```

To do this we bring an `nxpg.nix` script, with some patches for older pg
versions. This can be externalized later on as a flake to save on
compile time.