chore: speed up CI
Before each pg test took almost ~15 mins. See
https://github.com/supabase/supautils/actions/runs/13831763719/job/38697436630

Now each one takes 2 mins max. This is thanks to `cachix` and the centralization
of the pg tooling on https://github.com/steve-chavez/nxpg.
The location of this repo can be changed later.

This does change the way of testing from:

```
supautils-with-pg-13 make installcheck
```

To:

```
nxpg -v 13 test
```

But the speed up is worth it.