feat: no non-immutable funcs in check constraints
Problem:

A common failure mode of restores or other similar workflows is for a database to contain check constraints that are not immutable.
This is not supported by Postgres, but not flagged when such a constraint is created:
https://www.postgresql.org/docs/current/ddl-constraints.html#DDL-CONSTRAINTS-CHECK-CONSTRAINTS

Solution:

Forbid the use non-immutable functions on check constraint at CREATE or
ALTER time.

Implementation:

- [x] POC
- [ ] table level check constraint
- [ ] alter table statements
- [ ] expand on invalid functions instead of having an static array with
  `now` and `current_setting`